3d-gen/requirements.txt
PyYAML
cadquery==2.3.1
cadquery-ocp>=7.7.0a0,<7.8
ezdxf
multimethod>=1.7,<2.0
nlopt
nptyping==2.0.1
typish
casadi
path
